refactor: centralize vote update logic
