+++ events/messageReactionAdd.js
@@ -1,9 +1,8 @@
 // /events/messageReactionAdd.js
 
-const fs = require('fs');
-const path = require('path');
 const { handlePublicScan } = require('../lib/publicScanHandler');
 const { processFlaggedReview } = require('../lib/modReview');
+const { addVote } = require('../lib/voteUtils');
 
 module.exports = {
     name: 'messageReactionAdd',
@@ -29,25 +28,8 @@ module.exports = {
         const event = client.activeEvents && client.activeEvents.get(channelId);
         if (event) {
             const entry = event.entries.find(e => e.messageId === message.id);
-            if (entry) {
-                if (!['?', '❓', '✅', '❌', '🔁'].includes(emoji)) {
-                    if (entry.reactionUsers.has(user.id)) return;
-                    entry.reactionUsers.add(user.id);
-
-                    const oldPath = path.join(event.folder, entry.filename);
-                    const newScore = entry.reactionUsers.size;
-                    const newFilename = entry.filename.replace(/_rate\d+_/, `_rate${newScore}_`);
-                    const newPath = path.join(event.folder, newFilename);
-
-                    if (newFilename !== entry.filename) {
-                        try {
-                            fs.renameSync(oldPath, newPath);
-                            entry.filename = newFilename;
-                        } catch (err) {
-                            console.error('Failed to rename file:', err.message);
-                        }
-                    }
-                }
+            if (entry && !['?', '❓', '✅', '❌', '🔁'].includes(emoji)) {
+                addVote(entry, user.id, event);
             }
         }
